requirements.txt
Django==1.8.13
psutil==5.0.0

# Env
django-environ==0.4.0

# Celery: Message queue
celery==3.1.24
django-celery==3.1.17
celery-with-redis==3.0